# openapi/karma-v1.yaml
openapi: 3.1.0
info:
  title: Karma Ecosystem API
  version: 1.0.0
  summary: Payment intent, evidence verification, and risk alert APIs.
servers:
  - url: http://127.0.0.1:8811
security:
  - bearerAuth: []
paths:
...
  /v1/security/ops/alerts:
    get:
      summary: Get security operations alerts
      operationId: getSecurityOpsAlerts
      parameters:
        - in: query
          name: window_minutes
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1440
            default: 15
        - in: query
          name: failed_auth_threshold
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100000
            default: 10
        - in: query
          name: rate_limit_threshold
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100000
            default: 30
        - in: query
          name: private_runtime_error_threshold
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100000
            default: 5
        - in: query
          name: private_runtime_error_rate_threshold
          required: false
          schema:
            type: number
            minimum: 0.01
            maximum: 1
            default: 0.25
        - in: query
          name: private_runtime_min_requests
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100000
            default: 10
        - in: query
          name: dimension_limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 5
        - in: query
          name: alert_cooldown_minutes
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 1440
            default: 10
        - in: query
          name: failed_auth_threshold_overrides
          required: false
          schema:
            type: string
          description: "Comma-separated threshold overrides, e.g. '/v1/auth/token=5,group:auth=8'"
        - in: query
          name: rate_limit_threshold_overrides
          required: false
          schema:
            type: string
          description: "Comma-separated threshold overrides, e.g. '/v1/verify=20,group:verification=25'"
        - in: query
          name: private_runtime_error_threshold_overrides
          required: false
          schema:
            type: string
          description: "Comma-separated threshold overrides, e.g. '/v1/verify=3,group:verification=4'"
        - in: query
          name: private_runtime_error_rate_threshold_overrides
          required: false
          schema:
            type: string
          description: "Comma-separated ratio overrides, e.g. '/v1/verify=0.2,group:verification=0.3'"
        - in: query
          name: baseline_window_minutes
          required: false
          schema:
            type: integer
            minimum: 10
            maximum: 20160
            default: 1440
        - in: query
          name: baseline_drift_multiplier
          required: false
          schema:
            type: number
            minimum: 1.1
            maximum: 20
            default: 2.5
        - in: query
          name: baseline_min_sample_count
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100000
            default: 3
        - in: query
          name: baseline_capture_interval_minutes
          required: false
          schema:
            type: integer
            minimum: 0
            maximum: 1440
            default: 10
        - in: query
          name: apply_policy_center
          required: false
          schema:
            type: boolean
            default: true
        - in: query
          name: policy_id
          required: false
          schema:
            type: string
        - in: query
          name: policy_actor_id
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Security alert report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityOpsAlertReport'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /v1/security/policies:
    post:
      summary: Create security threshold policy version
      operationId: createSecurityThresholdPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecurityThresholdPolicyRequest'
      responses:
        '201':
          description: Security threshold policy created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityThresholdPolicy'
    get:
      summary: List security threshold policies
      operationId: listSecurityThresholdPolicies
      parameters:
        - in: query
          name: status
          required: false
          schema:
            $ref: '#/components/schemas/SecurityThresholdPolicyStatus'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        '200':
          description: Security threshold policy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecurityThresholdPolicy'
  /v1/security/policies/{policyId}:
    get:
      summary: Get security threshold policy
      operationId: getSecurityThresholdPolicy
      parameters:
        - in: path
          name: policyId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Security threshold policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityThresholdPolicy'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/security/policies/{policyId}/activate:
    post:
      summary: Promote policy to active (emergency override)
      operationId: activateSecurityThresholdPolicy
      parameters:
        - in: path
          name: policyId
          required: true
          schema:
            type: string
        - in: query
          name: emergency_override
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Activated policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityThresholdPolicy'
        '409':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/security/policies/{policyId}/candidate:
    post:
      summary: Set policy as rollout candidate (emergency override)
      operationId: setSecurityThresholdPolicyCandidate
      parameters:
        - in: path
          name: policyId
          required: true
          schema:
            type: string
        - in: query
          name: emergency_override
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SetSecurityThresholdPolicyCandidateRequest'
      responses:
        '200':
          description: Candidate policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityThresholdPolicy'
        '409':
          $ref: '#/components/responses/BadRequest'
  /v1/security/policies/rollback:
    post:
      summary: Roll back active policy (emergency override)
      operationId: rollbackSecurityThresholdPolicy
      parameters:
        - in: query
          name: emergency_override
          required: false
          schema:
            type: boolean
            default: false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RollbackSecurityThresholdPolicyRequest'
      responses:
        '200':
          description: Rolled-back active policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityThresholdPolicy'
        '409':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/security/policies/changes:
    post:
      summary: Create security policy change request
      operationId: createSecurityPolicyChangeRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecurityPolicyChangeRequest'
      responses:
        '201':
          description: Policy change request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityPolicyChangeRequest'
        '409':
          $ref: '#/components/responses/BadRequest'
    get:
      summary: List security policy change requests
      operationId: listSecurityPolicyChangeRequests
      parameters:
        - in: query
          name: status
          required: false
          schema:
            $ref: '#/components/schemas/SecurityPolicyChangeStatus'
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
      responses:
        '200':
          description: Policy change requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SecurityPolicyChangeRequest'
  /v1/security/policies/changes/{requestId}:
    get:
      summary: Get one security policy change request
      operationId: getSecurityPolicyChangeRequest
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Policy change request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityPolicyChangeRequest'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/security/policies/changes/{requestId}/review:
    post:
      summary: Review security policy change request
      operationId: reviewSecurityPolicyChangeRequest
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewSecurityPolicyChangeRequest'
      responses:
        '200':
          description: Updated policy change request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityPolicyChangeRequest'
        '409':
          $ref: '#/components/responses/BadRequest'
  /v1/security/policies/changes/{requestId}/apply:
    post:
      summary: Apply approved security policy change request
      operationId: applySecurityPolicyChangeRequest
      parameters:
        - in: path
          name: requestId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Applied policy change request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityPolicyChangeRequest'
        '409':
          $ref: '#/components/responses/BadRequest'
  /v1/security/policies/changes/dry-run:
    post:
      summary: Dry-run security policy change request
      operationId: dryRunSecurityPolicyChangeRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSecurityPolicyChangeRequest'
      responses:
        '200':
          description: Simulated impact report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SecurityPolicyDryRunResult'
        '409':
          $ref: '#/components/responses/BadRequest'
  /v1/security/runtime/safety-mode:
    get:
      summary: Get runtime safety mode state
      operationId: getRuntimeSafetyMode
      responses:
        '200':
          description: Runtime safety mode state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSafetyModeState'
    post:
      summary: Update runtime safety mode state
      operationId: updateRuntimeSafetyMode
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRuntimeSafetyModeRequest'
      responses:
        '200':
          description: Updated runtime safety mode state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSafetyModeState'
  /v1/security/runtime/anchor-audit:
    post:
      summary: Run runtime capacity-anchor audit
      operationId: runRuntimeAnchorAudit
      parameters:
        - in: query
          name: actor_id
          required: false
          schema:
            type: string
            default: system
      responses:
        '200':
          description: Anchor audit passed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeSafetyModeState'
        '503':
          description: Anchor audit failed and safety mode was enabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SecurityThresholdPolicyStatus:
      type: string
      enum: [draft, active, candidate, archived]
    SecurityThresholdPolicy:
      type: object
      properties:
        policy_id:
          type: string
        version:
          type: integer
        status:
          $ref: '#/components/schemas/SecurityThresholdPolicyStatus'
        rollout_percent:
          type: integer
        config:
          type: object
          additionalProperties: true
        note:
          type: string
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        activated_at:
          type: string
          format: date-time
        archived_at:
          type: string
          format: date-time
        parent_policy_id:
          type: string
      required:
        - policy_id
        - version
        - status
        - rollout_percent
        - config
        - created_at
    CreateSecurityThresholdPolicyRequest:
      type: object
      properties:
        config:
          type: object
          additionalProperties: true
        note:
          type: string
        created_by:
          type: string
        parent_policy_id:
          type: string
        rollout_percent:
          type: integer
          minimum: 0
          maximum: 100
          default: 100
      required: [config]
    SetSecurityThresholdPolicyCandidateRequest:
      type: object
      properties:
        rollout_percent:
          type: integer
          minimum: 1
          maximum: 99
          default: 10
      required: [rollout_percent]
    RollbackSecurityThresholdPolicyRequest:
      type: object
      properties:
        target_policy_id:
          type: string
...
    SecurityPolicyChangeStatus:
      type: string
      enum: [pending, approved, rejected, applied, cancelled]
...
    CreateSecurityPolicyChangeRequest:
      type: object
      properties:
        action:
          $ref: '#/components/schemas/SecurityPolicyChangeAction'
        target_policy_id:
          type: string
        target_rollback_policy_id:
          type: string
        rollout_percent:
          type: integer
          minimum: 1
          maximum: 99
        note:
          type: string
        requested_by:
          type: string
        required_approvals:
          type: integer
          minimum: 1
          maximum: 10
          default: 2
        dry_run_actor_id:
          type: string
      required: [action]
    ReviewSecurityPolicyChangeRequest:
      type: object
      properties:
        approver_id:
          type: string
        decision:
          $ref: '#/components/schemas/SecurityPolicyApprovalDecision'
        comment:
          type: string
      required: [approver_id, decision]
...
    UpdateRuntimeSafetyModeRequest:
      type: object
      properties:
        enabled:
          type: boolean
        reason:
          type: string
        actor_id:
          type: string
      required: [enabled]
...
    SecurityOpsAlertReport:
      type: object
      properties:
        window_minutes:
          type: integer
        summary:
          $ref: '#/components/schemas/SecurityOpsSummary'
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/SecurityOpsAlert'
        suppressed_alert_count:
          type: integer
        baseline:
          $ref: '#/components/schemas/SecurityOpsBaselineReference'
        policy_id:
          type: string
        policy_version:
          type: integer
        policy_status:
          $ref: '#/components/schemas/SecurityThresholdPolicyStatus'
        matched_candidate:
          type: boolean
        escalation:
          $ref: '#/components/schemas/SecurityOpsEscalationDecision'
        recommended_actions:
          type: array
          items:
            type: string
        generated_at:
          type: string
          format: date-time
      required:
        - window_minutes
        - summary
        - alerts
        - suppressed_alert_count
        - baseline
        - matched_candidate
        - escalation
        - recommended_actions
        - generated_at
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'